+++ src/main/java/roomescape/controller/page/LoginController.java
@@ -1,0 +1,13 @@
+package roomescape.controller.page;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class LoginController {
+
+    @GetMapping
+    public String home() {
+        return "login";
+    }
+}
