+++ src/main/java/roomescape/controller/rest/TimeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.rest;
 
 import java.time.LocalDate;
 import java.util.List;
