+++ src/main/java/roomescape/domain/Member.java
@@ -28,10 +28,6 @@ public class Member {
         return email;
     }
 
-    public String getPassword() {
-        return password;
-    }
-
     public Role getRole() {
         return role;
     }
