+++ src/main/java/roomescape/dto/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.dto.request;
+package roomescape.dto;
 
 import roomescape.domain.Member;
 import roomescape.domain.Role;
@@ -7,7 +7,6 @@ public record LoginMember(
     Long id,
     String name,
     String email,
-    String password,
     Role role
 ) {
 
@@ -16,7 +15,6 @@ public record LoginMember(
             user.getId(),
             user.getName(),
             user.getEmail(),
-            user.getPassword(),
             user.getRole()
         );
     }
